+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- *BREAKING* partiql-parser: Added a source location to `ParseError::UnexpectedEndOfInput`
 
 ### Added